+++ src/globals.h
@@ -130,6 +130,7 @@ const unsigned long c_bpmtag =     0x24240007;
 const unsigned long c_triggers_new =   0x24240008;
 const unsigned long c_midictrl =   0x24240010;
 const unsigned long c_mutegroups = 0x24240009; // not sure why we went to 10 above, this might need a different value
+const unsigned long c_resume = 0x24240011;
 
 
 const char c_font_6_12[] = "-*-fixed-medium-r-*--12-*-*-*-*-*-*";